refactor: 기존 tables api 함수 파일 내의 타입 코드 분리
