+++ src/app/api/tables.ts
@@ -1,30 +1,7 @@
 import { instance } from '@/lib/axios';
 import { getStoreId } from './store';
 import { OrderResponse } from './order';
-
-export type TablesResponse = {
-  content: {
-    id: number;
-    restaurantTableNumber: number;
-    displayStatus: string;
-    activeOrder?: {
-      orderId: number;
-      totalPrice: number;
-      orderItems: {
-        menuName: string;
-        quantity: number;
-        price: number;
-      }[];
-    } | null;
-  }[];
-  number: number;
-  totalPages: number;
-  totalElements: number;
-  size: number;
-  first: boolean;
-  last: boolean;
-  empty: boolean;
-};
+import { TablesResponse } from '@/types/table';
 
 export const PAGE_SIZE = 12;
